Destroying initial window to create a proper one in Windows.
I really need a bootstrap proccess to load OpenGL on Windows. It does
not work if I just use the previous window and load moder OpenGL
function. I really need to create a modern OpenGL context.

This is being done. After this I need to create again the window and
create the modern context.