Corrected the bugs that prevented execution on Windows.
Basically I should use the keyword '__stdcall' on all OpenGL function
pointers. Or the wrong calling convention would be used causing very
strange crashes.

I also declared using '#pragma' the libraries needed to compile in
Windows. Now it's not necessary to name explicitly the libraries that
should be linked with the code.