Written base code for keyboard support.
- But we still are wonkint o refactorate the code to turn keyboard and
mouse structs global variables. We will do this because anyway we do
not support more than one of such things and also we cannot create
more than one window. It was misleading to let the user manage mouse
and keyboard as if we could handle more complex options with more than
one of such things.